Rename battle inspection command file
