+++ commands/admin/cmd_battleinspect.py
